kf/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
beautifulsoup4==4.7.1
boltons==19.3.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
copyheaders==0.0.1
crypto==1.4.1
cryptography==2.2.2
cssselect==1.0.3
decorator==4.4.1
EasyProcess==0.2.7
face==19.1.2
Flask==1.0.2
glom==19.10.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
influxdb==5.2.1
itsdangerous==0.24
Jinja2==2.10
jsonpath==0.81
kafka==1.3.5
kafka-python==1.4.7
lxml==4.3.1
MarkupSafe==1.0
Naked==0.1.31
numpy==1.16.4
parsel==1.4.0
pika==1.0.0
Pillow==6.0.0
ply==3.11
psutil==5.5.1
py==1.8.0
py-zipkin==0.18.6
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
PyDispatcher==2.0.5
pyecharts-jupyter-installer==0.0.3
pyee==6.0.0
PyExecJS==1.5.1
pymongo==3.7.2
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyppeteer==0.0.25
pyquery==1.4.0
PySocks==1.7.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-logstash==0.4.6
pytz==2018.9
PyVirtualDisplay==0.2.4
PyYAML==5.1.2
queuelib==1.5.0
redis==2.10.6
requests==2.21.0
retry==0.9.2
Scrapy==1.5.0
scrapy-redis==0.6.8
selenium==3.12.0
service-identity==17.0.0
setup.py==0.1
shellescape==3.4.1
six==1.11.0
soupsieve==1.9.5
tenacity==5.0.3
threadpool==1.3.2
thriftpy2
tornado==5.1
tqdm==4.36.1
Twisted==18.4.0
urllib3==1.23
virtualenv==16.0.0
w3lib==1.19.0
websockets==8.0.2
Werkzeug==0.14.1
zope.interface==4.5.0
